+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+-   `JournalMode` dictates the journaling mode used by the database.
+
 ### Changed
 -   Restructure is now a SwiftPM project. All legacy build tools have been removed.
+-   The `Restructure` constructor takes a defaulted paramter for a journal mode.
 
 ## 1.0.0 - 2019-06-20
 ### Added